+++ src/js/core/switcher.js
@@ -1,5 +1,5 @@
 import Togglable from '../mixin/togglable';
-import {$$, addClass, attr, data, filter, getIndex, hasClass, index, isTouch, matches, queryAll, removeClass, toNodes, within} from 'uikit-util';
+import {$$, addClass, attr, data, filter, getIndex, hasClass, index, matches, queryAll, removeClass, toNodes, within} from 'uikit-util';
 
 export default {
 
@@ -83,10 +83,6 @@ export default {
             },
 
             handler(e) {
-                if (!isTouch(e)) {
-                    return;
-                }
-
                 e.preventDefault();
                 if (!window.getSelection().toString()) {
                     this.show(e.type === 'swipeLeft' ? 'next' : 'previous');